ref #85974 share exif and author credit rules across ingest paths
