+++ src/Domain/AuthorCleanPhrase/Cache/RegexAuthorCleanPhraseBuilder.php
@@ -17,10 +17,14 @@ final class RegexAuthorCleanPhraseBuilder extends AbstractAuthorCleanPhraseBuild
      */
     public function buildCache(AuthorCleanPhraseMode $mode, ExtSystem $extSystem): array
     {
-        if ($mode->isNot(AuthorCleanPhraseMode::Remove)) {
+        // Replace maps a matched phrase onto a replacement author through the named group the word
+        // builder emits; a hand-written regex has no such mapping. Remove and split only need the
+        // pattern itself, and a separator like "a dash with a space beside it" is not expressible
+        // as a quoted word.
+        if ($mode->is(AuthorCleanPhraseMode::Replace)) {
             throw new AuthorCleanPhraseException(
                 message: AuthorCleanPhraseException::ERROR_INVALID_MODE_AND_COMBINATION,
-                detail: 'Regex mode can be used only with remove mode. Mode ' . $mode->value . ' is not valid.'
+                detail: 'Regex type supports remove and split modes only. Mode ' . $mode->value . ' is not valid.'
             );
         }
